+++ src/configuration.rs
@@ -1,13 +1,17 @@
 use std::fs;
-use std::path::PathBuf;
-use std::sync::Mutex;
+//use std::fs::File;
+//use std::io::Write;
+//use std::path::{Path,PathBuf};
+use std::path::{PathBuf};
+use std::sync::RwLock;
 
 use config::*;
+use colored::*;
 
 use super::{ORGANIZATION, APPNAME, CONFIGNAME};
 
 lazy_static! {
-    static ref SETTINGS: Mutex<Config> = Mutex::new(Config::default());
+    static ref SETTINGS: RwLock<Config> = RwLock::new(Config::default());
 }
 
 pub fn config_dir() -> PathBuf {
@@ -17,7 +21,7 @@ pub fn config_dir() -> PathBuf {
 
 /// Reads the configuration from the file and the environment
 pub fn config_read() {
-    let mut settings = SETTINGS.lock().unwrap();
+    let mut settings = SETTINGS.write().unwrap();
 
     // The configuration file path is [OS dependant](https://crates.io/crates/directories)
     let config_path = config_dir();
@@ -53,15 +57,132 @@ pub fn config_read() {
     if let Err(e) = settings.merge(config::Environment::with_prefix(&APPNAME)) {
         println!("Failed parsing the environment settings:\n{}", e);
     }
+}
 
-    //println!("SETTINGS:\n--------\n{}, {:?}", APPNAME, settings); // DEBUG
+
+pub struct Lists {}
+
+impl Lists {
+
+    /// Checks if a list already exists
+    pub fn is(list: &str) -> bool {
+        let settings = SETTINGS.read().unwrap();
+
+        if let Ok(lists) = settings.get_table("lists") {
+            lists.contains_key(list)
+        } else {
+            false
+        }
+    }
+
+    /// Shows the crates contained in the list
+    pub fn show(list: &str, plain: bool) -> Option<String> {
+        let settings = SETTINGS.read().unwrap();
+
+        if let Ok(lists) = settings.get_table("lists") {
+            if let Some(value) = lists.get(list) {
+                if let Ok(array) = value.to_owned().into_array() {
+                    let mut crates_str = "".to_string();
+                    for value in array {
+                        if let Ok(v) = value.into_str() {
+                            if plain {
+                                crates_str = format!("{}, {}", crates_str, v.green());
+                            } else {
+                                crates_str = format!("{} {}", crates_str, v);
+                            }
+                        }
+                    }
+                    return Some(crates_str[1..].trim().to_string()); // remove leading comma
+                }
+            }
+        }
+        None
+    }
+
+    /// Returns the number of crates in a list
+    pub fn quantity(list: &str) -> usize {
+        let settings = SETTINGS.read().unwrap();
+        if let Ok(lists) = settings.get_table("lists") {
+            if let Some(value) = lists.get(list) {
+                if let Ok(array) = value.to_owned().into_array() {
+                    return array.len();
+                }
+            }
+        }
+        0
+    }
+
+
+    pub fn show_lists(recursive: bool) {
+        let settings = SETTINGS.read().unwrap();
+
+        if let Ok(lists) = settings.get_table("lists") {
+            if recursive {
+                for (list_name, value) in lists {
+                        if let Ok(array) = value.into_array() {
+                            println!("{}", list_name);
+                            for crate_name in array {
+                                println!("\t{}", crate_name);
+                            }
+                        }
+                }
+            } else {
+                //let lists_names: Vec<&String> = lists.keys().collect();
+                //println!("{:?}", lists_names[0]);
+
+                let mut lists_str = "".to_string();
+                for (list, _) in lists {
+                    lists_str = format!("{}, {} {}",
+                        lists_str, list.bright_green(),
+                        format!("({})", Self::quantity(&list)).cyan()
+                    );
+                }
+                println!("Your lists: {}", lists_str[1..].trim());
+            }
+        }
+
+    }
 }
 
 
 
+/*
+// Default configuration settings
+pub fn config_default_settings() {
+
+    let settings = SETTINGS.write().unwrap();
+
+    // TODO: save default values
+    //default_settings.set
+
+    let mut settings = SETTINGS.lock().unwrap();
+    if let Err(e) = settings.merge(default_conf)) {
+        println!("Error: failed to merge the default  file:\n{}", e);
+    }
+}
+*/
+
+/*
 pub fn config_write() {
 
+    let path = Path::new("test.toml");
+    let settings = SETTINGS.read().unwrap();
+
+    if let Ok(mut file) = fs::OpenOptions::new().write(true).open(path) {
+        file.set_len(0).unwrap();
+        for (key, value) in settings.collect().unwrap() {
+            file.write_all(
+                format!(
+                    "{}=\"{}\"\n",
+                    key,
+                    value.into_str().unwrap()
+                ).as_bytes()
+            ).unwrap();
+        }
+    }
+
 }
+*/